add etf/scr wsj commands
