+++ gamestonk_terminal/etf/screener/screener_controller.py
@@ -18,9 +18,10 @@ from gamestonk_terminal.helper_funcs import (
     parse_known_args_and_warn,
     try_except,
     system_clear,
+    check_positive,
 )
 from gamestonk_terminal.menu import session
-from gamestonk_terminal.etf.screener import screener_view
+from gamestonk_terminal.etf.screener import screener_view, wsj_view
 
 presets_path = os.path.join(os.path.abspath(os.path.dirname(__file__)), "presets/")
 
@@ -345,6 +346,76 @@ PRESET: {self.preset}
                 ascend=ns_parser.ascend,
                 export=ns_parser.export,
             )
+
+    @try_except
+    def call_gainers(self, other_args):
+        """Process gainers command"""
+        parser = argparse.ArgumentParser(
+            prog="gainers",
+            description="Displays top ETF/Mutual fund gainers from wsj.com/market-data",
+            add_help=False,
+        )
+        parser.add_argument(
+            "-l",
+            "--limit",
+            help="Limit of ETFs to display",
+            type=check_positive,
+            default=10,
+            dest="limit",
+        )
+        ns_parser = parse_known_args_and_warn(
+            parser, other_args, export_allowed=EXPORT_ONLY_RAW_DATA_ALLOWED
+        )
+        if other_args and "-" not in other_args[0][0]:
+            other_args.insert(0, "-l")
+        if ns_parser:
+            wsj_view.show_top_mover("gainers", ns_parser.limit, ns_parser.export)
+
+    def call_decliners(self, other_args):
+        """Process decliners command"""
+        parser = argparse.ArgumentParser(
+            prog="decliners",
+            description="Displays top ETF/Mutual fund decliners from wsj.com/market-data",
+            add_help=False,
+        )
+        parser.add_argument(
+            "-l",
+            "--limit",
+            help="Limit of ETFs to display",
+            type=check_positive,
+            default=10,
+            dest="limit",
+        )
+        ns_parser = parse_known_args_and_warn(
+            parser, other_args, export_allowed=EXPORT_ONLY_RAW_DATA_ALLOWED
+        )
+        if other_args and "-" not in other_args[0][0]:
+            other_args.insert(0, "-l")
+        if ns_parser:
+            wsj_view.show_top_mover("decliners", ns_parser.limit, ns_parser.export)
+
+    def call_active(self, other_args):
+        """Process gainers command"""
+        parser = argparse.ArgumentParser(
+            prog="active",
+            description="Displays most active ETF/Mutual funds from wsj.com/market-data",
+            add_help=False,
+        )
+        parser.add_argument(
+            "-l",
+            "--limit",
+            help="Limit of ETFs to display",
+            type=check_positive,
+            default=10,
+            dest="limit",
+        )
+        ns_parser = parse_known_args_and_warn(
+            parser, other_args, export_allowed=EXPORT_ONLY_RAW_DATA_ALLOWED
+        )
+        if other_args and "-" not in other_args[0][0]:
+            other_args.insert(0, "-l")
+        if ns_parser:
+            wsj_view.show_top_mover("active", ns_parser.limit, ns_parser.export)
 
 
 def menu(
